Complete Customer module (#6)
TO THE LIBRARY, this PR adds all endpoints available under the [customer API module](https://docs-v2.paystack.com/docs/api/customer/).

It also cleans up and refactors some of the implementation details for the plan module and adds proper error handling and a wrapper for handling errors.
It's a bit bulky due to some of the renaming changes of `subscription -> plan`, and consolidating the endpoints. I'm sorry in advance (lol).